Interactively handle missing image file
Sometimes, rather rarely actually, I receive an email with a missing
file link: there is an <img src...> tag but the file image file is
missing.

This patch makes org-msg ask the user if he/she wants to continue or
abort.  It allows the end user to fix the issue if he introduced it
like linking a non-existing image file in his reply for instance.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>