Fix color TODO HTML color string on Emacs > 25.1
Emacs 26 changed the `color-rgb-to-hex' function default behavior to
encode each color component with 16 bits instead of 8 bits (see commit
7b00e956 from Eli Zaretskii <eliz@gnu.org>).

This change also adds a `digits-per-component' optional parameter to
the `color-rgb-to-hex' function so that the old behavior can be
restored by passing the extra parameter value 2.

Because I would like to still be compatible with Emacs 25 for a while,
it left me with two options:

1. Add a test of the Emacs version and pass the extra parameter if
   necessary
2. Write my own implementation of `color-rgb-to-hex'

Since I don't want to have an Emacs version test in the org-msg module
and because the implementation of `color-rgb-to-hex' for the org-msg
module particular need fit in one line, I decided to own my own
version.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>
Signed-off-by: Boris Glavic <bglavic@iit.edu>