Support multiple greeting recipients
Adds a `org-msg-greeting-name-limit' variable to define the maximum
number of recipient first name for the greeting format.  If replying
to an email for which the 'To' field contains more than one recipient
and the `org-msg-greeting-fmt' contains a '%s' format, this variable
limits the number of recipient first name used as a replacement of the
'%s' format.  nil means unlimited.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>