Disable org-msg-mml-into-multipart-related with recursive mml
I fixed an issue in mml with two patches which have been recently
merged in the Emacs master branch.  These patches allow to get rid of
a hack based on placing an advice function on
`mml-expand-html-into-multipart-related'.

Before these two patches the `mml-generate-mime' function needed the
"text/html" element to be the only once of an mml message to
automatically included the images as a multipart `related' element.
To be able to keep this feature and also be able to attach raw files
to a an email, I created an advice on the
`mml-expand-html-into-multipart-related' function to create a multipart
`mixed' element and include all the attachments.

This patch detects if the presence of these two patches.  If they are
present if does not place an advice on
`mml-expand-html-into-multipart-related' and use mml tags to structure
the email.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>