Make `org-msg-prepare-to-send' more robust
Tim Quelch reported in #58 that the notmuch MUA directly and
indirectly runs the `message-send-hook'. This makes the message
converted to HTML twice, causing raw HTML to be included in the sent
message.

The best way to fix this is certainly in motmuch. Tim reported this
issue to the notmuch developers through their mailing list (see
"emacs: `message-send-hook` being called twice").

There are many ways to work-around this issue as removing
`org-msg-prepare-to-send' from the `message-send-hook' ...  But it
appears to me that instead of creating a workaround we could make the
`org-msg-prepare-to-send' more robust.  The goal of this function is
to convert the OrgMode content of the org-msg-edit-mode buffer to
MML/HTML content.  Instead of assuming that the content is OrgMode,
this function will now use a text property to know if the content is
OrgMode or MML.

When the `org-msg-prepare-to-send' is called more than once, it prints
a warning message. It gives a chance to the user to detect and report
a potential issue with a backend implementation.

Reported-by: Tim Quelch <tim@tquelch.com>
Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>