Make `message-exit-actions' permanent local
This is a follow-up for the #58. The `message-exit-actions' is
supposed to be a local variable.  It loses this property when
`org-msg-edit-mode' major mode is loaded. This variable contains
expressions to restore the windows configuration preceding the
creation of a message buffer. Being global it keeps accumulating
expressions to get back to the windows configuration preceding the
first message buffer creation.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>