Update the charset of the original email
This patch addresses #80.

When replying to an email, the HTML content is loaded in Emacs and the
strings are converted to the Emacs internal representation. Therefore
the strings of the original email are no more subject to the charset
stated in the meta tags.

It is not clear what the right charset is once the content goes
through the MML encoding.  For now, we assume that the current charset
is the `org-html-coding-system' charset which by default is utf-8.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>