Make org babel generate absolute result file path
When composing an email, the `default-directory' and the
`buffer-file-name' of are different. It makes org babel generate
cumbersome file path if the `org-link-file-path-type' is set the
`adaptive' which is the default value. These unexpected file paths can
prevent a proper rendering in a web browser.

As suggested by Alexandre Duret-Lutz, `org-msg-edit-mode' sets the buffer local
`org-link-file-path-type' variable to `absolute' making org babel
generating absolute path.

This patch addresses #83.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>