Handle alternatives in org-msg-preview
If the current OrgMsg buffer does not request the `html' alternative,
it does not make sense for `org-msg-preview' to export as html and
open the web browser.

This patch makes `org-msg-preview' export for the last alternative of
the "alternatives" property. Indeed, the alternatives should be listed
in increasing order of preference.  If this alternative is `html' it
calls the `browse-url' function to display the exported mail in a web
browser. For all other alternatives, it displays the exported result
in a buffer.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>