Prevent text re-arrangement by `org-ascii-export-as-ascii'
The `org-ascii-export-as-ascii' uses the `fill-region' function which
among other operations breaks lines and suppresses duplicated
white-spaces.  This is not a desirable behavior for text
exportation. Indeed, `fill-region' can lead to best practices
violation by improperly breaking quoted text or as reported by
Michael-David Fiszer in #89, signature misinterpretation. Indeed,
`fill-region' turns '^--[[:blank:]]$' into '^--$'. But '^--$' is the
signature marker as per Usenet Best Practice 3.1.2.1
[https://tools.ietf.org/html/draft-ietf-usefor-useage-01#section-3.1.2.1].

In order to prevent any call to `fill-region' during the execution of
`org-ascii-export-as-ascii', this patch temporarily binds the
`fill-region' function to the `ignore' function.

This patch also prevents the inner margin to be added during the
exportation.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>