Disable OrgMsg on messages containing MML tags
This patch addresses #92 as well as the mu4e issue djcb/mu#1956 (once
djcb/mu#1967 is merged for the latter).

This patch is a port of the original suggestion from Christophe
Troestler 4c46b0f8 (#95).

OrgMsg does not support MML tags and this patch makes it not activate
itself when some MML tags are present.

Typically, if you have `mu4e-compose-forward-as-attachment' set to t,
on a mail forward, mu4e includes MML tags in the email buffer. OrgMsg
does not know how to handle them and this is why it is better to not
activate OrgMsg in such a situation.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>
Signed-off-by: Christophe Troestler <Christophe.Troestler@umons.ac.be>