Allow more flexibility in the specification of alternatives
This patch addresses request #90.

The end-user may want to automatically set the current reply
alternatives list based on various factors such as `To:` mail address,
`Cc:' mail addresses, `Subject:', even a regular expression match on
the entire buffer.

To satisfy such a wide range of possibilities, this patch makes the
`org-msg-post-setup' function call a new function named
`org-msg-get-alternatives'. This new function is responsible for
returning the list of alternatives for the current message buffer. By
advising this function, it becomes possible to implement any desired
behavior.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>