Add support to reply to text/plain emails
This patch implements #89 and also addresses a comment of #83.

As it has been fairly pointed to me recently, Org Mode can be very
powerful to compose an email even if the result is going to be ASCII
as it makes it easy to create tables, perform arithmetic operations,
quote code or create diagram on the fly.

This patch modified the `org-msg-default-alternatives' variable
specification to offer a more flexible customization of the list of
alternatives for new emails, or replies to text/html and text/plain
emails.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>