Allow more flexibility in the specification of composition parameter
This patch addresses #104.

It would be impossible for OrgMsg to offer customization variables to
accommodate all the combinations of desirable behaviors. One could
want to change its signature when replying as plain text only. In
addition, email can be used to interact with machines and their
protocol can have even more exotic requirements.

In order to be able to dynamically change the posting style, greeting
format or signature, these parameters are now returned by a new
`org-msg-composition-parameters' function which can be advised by the
end-user to implement specific behaviors.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>