Introduce the org-msg-setup-on-message-mail customization variable
It should not be necessary to advise the message-mail function, every
MUA has its own composition function or hook.

If for some reason, you still need to make `message-mail' to be
followed by `org-msg-post-setup', you should set
`org-msg-setup-on-message-mail' to t.

Note that I am planning to remove it at some points unless I hear that
this is needed. Therefor if you really need it raise your concern over
email.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>