Add support to insert a mailto link to a recipient
It has become more and more popular in the last few years to refer to
a recipient of an email by adding a mailto link.  The
`org-msg-insert-recipient-mailto' function can be use to insert such a
mailto link. It is bound to "C-c @" in place of the default
`org-mark-subtree' org-mode function.

This commit addresses #179.

Signed-off-by: Jeremy Compostella <jeremy.compostella@gmail.com>
Suggested-by: Alessio Bolognino (molok)