mu4e: Fix failure to start `org-msg-edit-mode` when editing drafts
Editing a draft message with a subject fails. In `org-msg-post-setup`
when calling `(org-msg-mua-call 'article-htmlp)` it fails with "No
message at point" from `mu4e-message-at-point`.

To solve this we advice `mu4e-compose-edit` and set a flag that we can
check in `org-msg-post-setup`. This check is setup using a MUA call,
although this is only implemented for mu4e. If this is need for other
backends they need to implement the predicate and clear draft flag
functions.