Fix crash with new Emacs version by ignoring browse-url-of-file
The latest version of Emacs introduced changes in the behavior of
`browse-url-of-file`, which led to unexpected crashes when handling HTML
articles in `org-msg.el`. To mitigate this issue, added
`browse-url-of-file` to be ignored alongside
`browse-url-browser-function` during the processing of Gnus
articles. This ensures consistent behavior when extracting HTML content
from articles.

Signed-off-by: Jeremy Compostella <jeremy.compostella@intel.com>